requirements.txt
fastapi
uvicorn
groq
supabase
pydantic
asyncio
python-dotenv 